Add ChallengesList and ChallengesListHeader components for improved challenge display
