+++ src/app/page.tsx
@@ -1,36 +1,17 @@
 "use client";
 
-import ChallengeCard from "@/components/cards/ChallengeCard";
-import { Loader, Text, Title } from "@mantine/core";
 import WelcomeBar from "@/components/groups/welcome-bar/WelcomeBar";
-import { useAllChallenges } from "@/classes/Challenge/AllChallengesHook";
+import ChallengesList from "@/components/groups/challenge/ChallengesList";
+import ChallengesListHeader from "@/components/groups/challenge/ChallengesListHeader";
 
 export default function Home() {
-    const challenges = useAllChallenges();
-
     return (
-        <div className={`flex flex-col space-y-10`}>
+        <div className={`flex flex-col space-y-10 w-full`}>
             <WelcomeBar />
 
             <div className={`flex flex-col space-y-5`}>
-                <div>
-                    <Title order={1} className={`text-gray-700`}>
-                        Challenges
-                    </Title>
-                    <Text size="md" className={`text-gray-500`}>
-                        Complete these challenges to earn points!
-                    </Text>
-                </div>
-
-                {challenges !== null && challenges.length !== 0 && (
-                    <div className={`grid grid-cols-3 gap-8 justify-start`}>
-                        {challenges.map((challenge, index) => (
-                            <ChallengeCard key={index} challenge={challenge} />
-                        ))}
-                    </div>
-                )}
-
-                {challenges === null && <Loader />}
+                <ChallengesListHeader />
+                <ChallengesList />
             </div>
         </div>
     );
